Refactor game logic and state into new Game Class
The game's state and functionality should be encapsulated in its own
class.  This allows a common game-engine design pattern in which
different "game modes" can exist within the same project, and a "driver"
program is responsible for transitioning between different modes.  This
also cleanly separates the "client" (the program responsible for
managing the window, input, etc.) from the actual game.